==> Section_5_neuroectoderm/step6_Key_TFs.R <==

##################################
### Section - 5, Neuroectoderm ###
##################################

###################################################################################
### Plotting gene expression over time for top TFs for each spinal interneurons ###
###################################################################################

source("JAX_help_code.R")
source("JAX_color_code.R")
work_path = "./"

example_i = "Neurons"; print(example_i)
pd = readRDS(paste0(work_path, example_i, "_adata_scale.obs.rds"))
rownames(pd) = as.vector(pd$cell_id)

mouse_gene_sub = mouse_gene[(mouse_gene$gene_type %in% c('protein_coding', 'pseudogene', 'lincRNA')) & mouse_gene$chr %in% paste0("chr", c(1:19, "M")),]
gene_count = doExtractData(pd, mouse_gene_sub)
obj = CreateSeuratObject(gene_count, pd)

obj$keep = obj$neurons_sub_clustering %in% c("Spinal dI1 interneurons",
                                             "Spinal dI2 interneurons",
                                             "Spinal dI3 interneurons",
                                             "Spinal dI4 interneurons",
                                             "Spinal dI5 interneurons",
                                             "Spinal dI6 interneurons",
                                             "Spinal V0 interneurons",
                                             "Spinal V1 interneurons",
                                             "Spinal V2a interneurons",
                                             "Spinal V2b interneurons",
                                             "Spinal V3 interneurons")
obj = subset(obj,subset = keep)
Idents(obj) = as.vector(obj$neurons_sub_clustering)
obj = subset(obj, downsample = 10000)
obj = NormalizeData(obj, normalization.method = "LogNormalize", scale.factor = 10000)
data = GetAssayData(obj, slot = "data")

result = read.csv(paste0(work_path, "Neurons_interneurons_top_TFs.csv"))
celltype_list = names(table(result$cluster))

celltype_list = paste0("Spinal dI", c(1:5), " interneurons")

early_list = NULL

### Extended Data Fig. 10l

for(celltype_i in celltype_list){
    
    result_sub = result %>% filter(cluster == celltype_i)
    
    day_list_sub = names(table(as.vector(obj$day[obj$celltype_sub_clustering == celltype_i])))
    day_list_sub = as.vector(day_list[day_list %in% day_list_sub])
    
    exp = NULL
    for(day_i in day_list_sub){
        print(day_i)
        exp = cbind(exp, Matrix::rowMeans(data[rownames(data) %in% as.vector(result_sub$gene_ID),obj$celltype_sub_clustering == celltype_i &
                                                   obj$day == day_i, drop=FALSE]))
    }
    
    exp = exp[as.vector(result_sub$gene_ID),]
    rownames(exp) = as.vector(result_sub$gene_short_name)
    colnames(exp) = day_list_sub
    
    exp_row_max = sort(apply(exp, 1, which.max))
    exp = exp[names(exp_row_max),]
    
    early_list = rbind(early_list, data.frame(celltype = celltype_i,
                                              gene = names(exp_row_max)[exp_row_max == 1]))
    
    name = gsub("Spinal ", "", celltype_i)
    name = gsub(" interneurons", "", name)
    
    Colors=rev(brewer.pal(11,"Spectral"))
    Colors=colorRampPalette(Colors)(120)
    pdf(paste0(work_path, name, "_heatmap.pdf"), 8, 8)
    heatmap.2(as.matrix(exp), 
              col=Colors, 
              scale="none", 
              Rowv = F, 
              Colv = F, 
              key=T, 
              density.info="none", 
              trace="none", 
              cexRow = 0.5, 
              cexCol = 1,
              margins = c(5,5))
    dev.off()
    
}

###############################################################################################################
### Gene expression for selected TFs, across progenitor cells of dI1-5 from the neuroectodermal territories ###
###############################################################################################################

### First, identifying the progenitor cells for dI1-5 based on MNNs
### Some temporary results are necessary to run step4_Mapping_neuroectoderm_derivatives.R first.

source("JAX_help_code.R")
source("JAX_color_code.R")
work_path = "./"

example_i = "Neuroectoderm_derivative"

pd = readRDS(paste0(work_path, example_i, "_adata_scale.obs.rds"))

### Neuroectoderm_derivative.MNN.rds was generated by step4_Mapping_neuroectoderm_derivatives.R
dat = readRDS(paste0(work_path, example_i, ".MNN.rds"))
pd_1 = dat[["pd_1"]]
pd_2 = dat[["pd_2"]]
nn_matrix = dat[["nn_matrix"]]
nn_matrix_2 = dat[["nn_matrix_2"]]

k.param = 10

resultA = NULL
for(i in 1:k.param){
    print(i)
    resultA = cbind(resultA, as.vector(pd_2$cell_id)[as.vector(nn_matrix[,i])])
}
datA = data.frame(A = rep(pd_1$cell_id, k.param),
                  B = c(resultA), 
                  value = 1, stringsAsFactors = F)

resultB = NULL
for(i in 1:k.param){
    print(i)
    resultB = cbind(resultB, as.vector(pd_1$cell_id)[as.vector(nn_matrix_2[,i])])
}
datB = data.frame(A = c(resultB),
                  B = rep(pd_2$cell_id, k.param), 
                  value = 2, stringsAsFactors = F)


dat = datB %>% left_join(datA, by = c("A" = "A", "B" = "B")) %>%
    filter(!is.na(value.y)) %>% select(A, B) %>%
    left_join(pd_1 %>% mutate(A = cell_id) %>% select(A, celltype_A = celltype), by = "A") %>%
    left_join(pd_2 %>% mutate(B = cell_id) %>% select(B, celltype_B = celltype), by = "B")

row_names_order = c("Spinal dI1 interneurons",
                    "Spinal dI2 interneurons",
                    "Spinal dI3 interneurons",
                    "Spinal dI4 interneurons",
                    "Spinal dI5 interneurons")

### a progenitor cell from backbone is mutually exclusive to a given neuron type
dat_m = dat %>% filter(celltype_A %in% row_names_order) %>% group_by(B, celltype_A) %>% tally() %>%
    group_by(B) %>% slice_max(order_by = n, n = 1, with_ties = F) %>%
    rename(cell_id = B, celltype_derive = celltype_A, freq = n)

### a progenitor cell from backbone is not required to be exclusive to a given neuron type
dat_n = dat %>% filter(celltype_A %in% row_names_order) %>%
    rename(cell_id = B, celltype_derive = celltype_A)

pd_sub = pd %>% filter(cell_id %in% as.vector(dat_m$cell_id))

### For each cell, we only retain protein-coding genes, lincRNA genes and pseudogenes
mouse_gene_sub = mouse_gene[(mouse_gene$gene_type %in% c('protein_coding', 'pseudogene', 'lincRNA')) & mouse_gene$chr %in% paste0("chr", c(1:19, "M")),]
gene_count = doExtractData(pd_sub, mouse_gene_sub)

saveRDS(list(gene_count = gene_count, pd_sub = pd_sub,
             dat_m = dat_m, dat_n = dat_n), paste0(work_path, "IN_backbone_prog.rds"))

### Neuroectoderm_derivative.MNN.rds was generated by step4_Mapping_neuroectoderm_derivatives.R
dat = readRDS(paste0(work_path, example_i, ".MNN.rds"))
pd_1 = dat[["pd_1"]]
pd_sub = pd_1[pd_1$celltype_update %in% row_names_order,]

### For each cell, we only retain protein-coding genes, lincRNA genes and pseudogenes
mouse_gene_sub = mouse_gene[(mouse_gene$gene_type %in% c('protein_coding', 'pseudogene', 'lincRNA')) & mouse_gene$chr %in% paste0("chr", c(1:19, "M")),]
gene_count = doExtractData(pd_sub, mouse_gene_sub)

obj_early = CreateSeuratObject(gene_count, meta.data = pd_sub)
obj_early$group = "early"
saveRDS(obj_early, paste0(work_path, "IN_early_cells.rds"))


### Second, plot gene expression for top TFs in the progenitors for each spinal interneurons

source("JAX_help_code.R")
source("JAX_color_code.R")
work_path = "./"

row_names_order = c("Spinal dI1 interneurons",
                    "Spinal dI2 interneurons",
                    "Spinal dI3 interneurons",
                    "Spinal dI4 interneurons",
                    "Spinal dI5 interneurons")

res_overlap = read.csv(paste0(work_path, "Neurons_interneurons_overlap_TFs.csv"))
res_x = res_overlap %>% left_join(mouse_gene[,c("gene_ID","gene_short_name")], by = "gene_short_name")
res_overlap$gene_ID = as.vector(res_x$gene_ID)

dat = readRDS(paste0(work_path, "IN_backbone_prog.rds"))

dat_n = dat[["dat_n"]]

obj = CreateSeuratObject(dat[["gene_count"]], meta.data = dat[["pd_sub"]])
obj = NormalizeData(obj, normalization.method = "LogNormalize", scale.factor = 10000)
data = GetAssayData(obj, slot = "data")

exp = NULL
for(celltype_i in row_names_order){
    cell_include = as.vector(dat_n %>% filter(celltype_derive == celltype_i) %>% pull(cell_id))
    exp = cbind(exp, Matrix::rowMeans(data[rownames(data) %in% as.vector(res_overlap$gene_ID),
                                           colnames(data) %in% cell_include, 
                                           drop=F]))
}

colnames(exp) = row_names_order
exp = exp[as.vector(res_overlap$gene_ID),]
rownames(exp) = as.vector(res_overlap$gene_short_name)

### Extended Data Fig. 10m (left)

Colors=rev(brewer.pal(11,"Spectral"))
Colors=colorRampPalette(Colors)(120)
pdf(paste0(work_path, "progenitor_heatmap.pdf"), 8, 8)
heatmap.2(as.matrix(t(exp)), 
          col=Colors, 
          scale="col", 
          Rowv = F, 
          Colv = F, 
          key=T, 
          density.info="none", 
          trace="none", 
          cexRow = 0.5, 
          cexCol = 1,
          margins = c(5,5))
dev.off()


### Third, plot gene expression over time bins for each spinal interneurons

source("JAX_help_code.R")
source("JAX_color_code.R")
work_path = "./"

example_i = "Neurons"; print(example_i)
pd = readRDS(paste0(work_path, example_i, "_adata_scale.obs.rds"))
rownames(pd) = as.vector(pd$cell_id)

mouse_gene_sub = mouse_gene[(mouse_gene$gene_type %in% c('protein_coding', 'pseudogene', 'lincRNA')) & mouse_gene$chr %in% paste0("chr", c(1:19, "M")),]
gene_count = doExtractData(pd, mouse_gene_sub)
obj = CreateSeuratObject(gene_count, pd)

obj$keep = obj$neurons_sub_clustering %in% c("Spinal dI1 interneurons",
                                             "Spinal dI2 interneurons",
                                             "Spinal dI3 interneurons",
                                             "Spinal dI4 interneurons",
                                             "Spinal dI5 interneurons",
                                             "Spinal dI6 interneurons",
                                             "Spinal V0 interneurons",
                                             "Spinal V1 interneurons",
                                             "Spinal V2a interneurons",
                                             "Spinal V2b interneurons",
                                             "Spinal V3 interneurons")
obj = subset(obj,subset = keep)
Idents(obj) = as.vector(obj$neurons_sub_clustering)
obj = subset(obj, downsample = 10000)

row_names_order = c("Spinal dI1 interneurons",
                    "Spinal dI2 interneurons",
                    "Spinal dI3 interneurons",
                    "Spinal dI4 interneurons",
                    "Spinal dI5 interneurons")

res_overlap = read.csv(paste0(work_path, "Neurons_interneurons_overlap_TFs.csv"))
res_x = res_overlap %>% left_join(mouse_gene[,c("gene_ID","gene_short_name")], by = "gene_short_name")
res_overlap$gene_ID = as.vector(res_x$gene_ID)

obj_early = readRDS(paste0(work_path, "IN_early_cells.rds"))
obj_early = NormalizeData(obj_early, normalization.method = "LogNormalize", scale.factor = 10000)

obj$celltype_update = as.vector(obj$neurons_sub_clustering)
obj$keep = obj$celltype_update %in% row_names_order & 
    (!obj$cell_id %in% obj_early$cell_id)
obj_sub = subset(obj, subset = keep)
obj_sub$group = "late"
obj_sub = NormalizeData(obj_sub, normalization.method = "LogNormalize", scale.factor = 10000)

dat_early = GetAssayData(obj_early, slot = "data")[as.vector(res_overlap$gene_ID),]
dat_late = GetAssayData(obj_sub, slot = "data")[as.vector(res_overlap$gene_ID),]
pd_sub = data.frame(obj_sub[[]])

pd_all = readRDS(paste0(work_path, "df_cell.rds"))
pd_x = pd_sub %>% left_join(pd_all[,c("cell_id","somite_count")], by = "cell_id")
pd_sub$somite_count = as.vector(pd_x$somite_count)

day_x = data.frame(somite_count = paste0(c(27:45), " somites"),
                   day_value = c(1:length(c(27:45))))
day_y = pd_sub %>% left_join(day_x, by = "somite_count")
pd_sub$day_value = as.vector(day_y$day_value)

exp = NULL
for(i in 1:nrow(res_overlap)){
    print(i)
    celltype_i = as.vector(res_overlap$cluster)[i]
    gene_i = as.vector(res_overlap$gene_ID)[i]
    
    exp_i = NULL
    exp_i = c(exp_i, mean(as.vector(dat_early[gene_i, obj_early$celltype_update == celltype_i])))
    
    pd_i = pd_sub[pd_sub$celltype_update == celltype_i,]
    print(table(pd_i$day_value))
    pd_i = pd_i[order(pd_i$day_value),]
    bin_size = floor(nrow(pd_i)/20)+1
    for(j in 1:20){
        
        bin_coor = ((j-1)*bin_size + 1):(j*bin_size)
        
        if(j == 20){
            bin_coor = ((j-1)*bin_size + 1):nrow(pd_i)
        }
        
        pd_j = pd_i[bin_coor,]
        exp_i = c(exp_i, mean(as.vector(dat_late[gene_i, colnames(dat_late) %in% as.vector(pd_j$cell_id)])))
    }
    exp = cbind(exp, exp_i)
}

rownames(exp) = c("Earliest 500", paste0("bin_", 1:20))
colnames(exp) = as.vector(res_overlap$gene_short_name)

### Extended Data Fig. 10m (right)

Colors=rev(brewer.pal(11,"Spectral"))
Colors=colorRampPalette(Colors)(120)
pdf(paste0(work_path, "timepoint_heatmap.pdf"), 8, 8)
heatmap.2(as.matrix(exp), 
          col=Colors, 
          scale="col", 
          Rowv = F, 
          Colv = F, 
          key=T, 
          density.info="none", 
          trace="none", 
          cexRow = 0.5, 
          cexCol = 1,
          margins = c(5,5))
dev.off()
